Defines request for image data

ee_data_image() and ee_data_collection() are used to define the requested earth enigne data for the ee_grab() function.

ee_data_image(datasetID = "CGIAR/SRTM90_V4", spatialReducer = "mean",
  scale = 3000, bandSelection = NULL)

Arguments

datasetID
string that specifies the dataset in Earth Engine. The dataset ID can be found in the snippet section of the dataset in the Earth Engine Data Catalog.

spatialReducer
string that specifies the spatial aggregation of the data within the polygons of the targetArea. The spatial reducer can be one of "mean", "median", "min", "max", "mode"

scale
integer that controls the scale of analysis in Earth Engine. The scale controls the resolution of the data in which the computations are performed. In Earth Engine data is ingested at multiple scales, in an image pyramid. When you use an image, Earth Engine chooses a level of the pyramid with the closest scale less than or equal to the scale specified by your scale argument and resamples (using nearest neighbour by default) as necessary. For more information use the Link.

bandSelection
string or a vector of strings of bands names to select from the requested dataset. By default bandSelection is set to NULL and all bands of the dataset are used.

Value

object of class list that defines the data request for ee_grab().

Image and Image Collections in Earth Engine

In Earth Engine raster data is stored as an Image object. Images are composed of one or more bands and each band has its own name, data type, scale, mask and projection. A time series or stack of Images is stored as an Image Collection. To request data from an Image use ee_data_image() to define the request. To request data from a time series of Images stored in an Image Collection use ee_data_collection() instead.