Install dependencies and run authentications

ee_grab_install() installs the required dependencies and guides the user through the authentication processes to activate the different API's.

ee_grab_install(clean_credentials = T, clean_environment = F)

Arguments

clean_credentials
logical, if True already present credential are deleted recreated by a reauthenticate. Default is set to True. This argument is used for development and not meant to be changed by the user.

clean_environment
logical, if True already installed environments are deleted to be reinstalled again. Default is set to False. This argument is used for development and not meant to be changed by the user.

Installation of Dependencies

To encapsulate the dependencies from the users system, and at the same time simplify the installation the earthEngineGrabR, uses a conda environment. By running ee_grab_install() first the conda envritonemt "earthEngineGrabR" is created. Further all dependencies are installed inside the "earthEngineGrabR" envrironment.

Authentication of API's

The earthEngineGrabR connects to 3 Google API's:

Google Fusion Table API for uploading data.

Google Earth Engine API for data aquisition and processing.

Google Drive API for data download.

To authenticate to the API's the user has to log in with his google account and allow the API to access data on googles servers on the user's behalf. To simplify this procedure the ee_grab_install function successively opens a browser window to log into the Google account. If the Google account is verified and the permission is granted, the user is directed to an authentification token. This token is manually copied and pasted into the R console, which creates persistent credentials. This process is repeated for each API. If the function runs successfully, all needed credentials are stored for further sessions and there should be no need for further authentification.